requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.0
PyJWT==2.8.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.0.1
psycopg2-binary==2.9.7
twilio==8.9.1
argon2-cffi==23.1.0
email-validator==2.1.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0 